# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Athena Pay API
  description: |
    # Athena Pay - API Documentation

    API RESTful da plataforma financeira digital Athena Pay.

    ## Autenticacao

    Todas as requisicoes requerem autenticacao via Bearer Token (JWT) obtido atraves do Keycloak.

    ```
    Authorization: Bearer <access_token>
    ```

    ## Rate Limiting

    | Endpoint | Limite |
    |----------|--------|
    | /auth/* | 10 req/min |
    | /pix/* | 100 req/min |
    | /accounts/* | 200 req/min |
    | Outros | 500 req/min |

    ## Ambientes

    | Ambiente | URL |
    |----------|-----|
    | Producao | https://api.athenapay.com.br |
    | Sandbox | https://sandbox.athenapay.com.br |
    | Desenvolvimento | http://localhost:8000 |

  version: 2.0.0
  contact:
    name: Athena Pay Support
    email: api@athenapay.com.br
    url: https://developers.athenapay.com.br
  license:
    name: Proprietary
    url: https://athenapay.com.br/terms
servers:
  - url: https://api.athenapay.com.br/v1
    description: Producao
  - url: https://sandbox.athenapay.com.br/v1
    description: Sandbox
  - url: http://localhost:8000/api
    description: Desenvolvimento
...
paths:
...
  /pix/keys:
    get:
      tags: [PIX]
      summary: Listar chaves PIX
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de chaves
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PIXKey'

    post:
      tags: [PIX]
      summary: Cadastrar chave PIX
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PIXKeyRequest'
      responses:
        '201':
          description: Chave cadastrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PIXKey'
        '409':
          description: Chave ja cadastrada
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    PIXKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [CPF, CNPJ, EMAIL, PHONE, EVP]
        value:
          type: string
        status:
          type: string
          enum: [active, pending, inactive]
        createdAt:
          type: string
          format: date-time
    PIXKeyRequest:
      type: object
      required: [type, value, accountId]
      properties:
        type:
          type: string
          enum: [CPF, CNPJ, EMAIL, PHONE, EVP]
        value:
          type: string
        accountId:
          type: string
          format: uuid
...
security:
  - bearerAuth: []